Don't delete packages dir unless it already exists

+++ build-with-casper.cmd
@@ -2,6 +2,6 @@
 
 set myDir=%~dp0
 
-rmdir /S /Q %myDir%out || goto :eof
+if exist %myDir%out rmdir /S /Q %myDir%out || goto :eof
 set TargetFrameworkVersion=v4.5.1
 call %myDir%casper %* || goto :eof
